requirements.txt
langchain==0.3.25
langchain-openai==0.3.18
langchain-pinecone==0.2.8
langchain-core==0.3.65
pinecone-client==6.0.0
openai==1.86.0
python-dotenv==1.1.0
gradio==5.32.1
pandas==2.2.3
matplotlib==3.10.3